requirements.txt
streamlit
random
numpy
google.generativeai
python-dotenv
time